added to HomeController, added DressingDao

+++ src/main/java/com/tessasanders/dressingselector/models/data/DressingDao.java
@@ -7,6 +7,6 @@ import javax.transaction.Transactional;
 
 @Repository
 @Transactional
-public interface DressingDao extends CrudRepository<Dressing, Integer> {
+public interface DressingDao extends CrudRepository<Dressing, Long> {
 
 }
